refactor: Reorganize test files into logical subfolders
- Create tests/providers/ subfolder for provider-specific tests
- Move test_backend_switching.py to tests/integration/
- Move test_cross_entry_detection.py to tests/unit/
- Move test_marian_client.py to tests/providers/
- Update backend switching tests to work with plugin system
- Fix path issues in moved test files
- Update error message expectations for plugin system validation

Test structure now:
- tests/unit/ - Unit tests for algorithms and utilities
- tests/integration/ - Integration tests for end-to-end flows
- tests/providers/ - Provider-specific implementation tests
- tests/debug/ - Debug scripts
- tests/data/ - Test data files
